Move `ProfileAttributes` to the `attributes` package
